Pass all unittest in MSVC Debug
The standard library of microsoft contains in the default debug configuration
a iterator-check. This check is not compatible with the iterator design choice
of xtensor. The check should be disabled for MSVC Debug.

The check is easily disabled with the define "_ITERATOR_DEBUG_LEVEL=0"
A new option "DISABLE_MSVC_ITERATOR_CHECK" is created to control this behaviour.
The default value of the option is "ON".

With this solution, the unittest can also be executed in debug mode on a windows
machine.